refactory user

+++ src/controllers/userController.js
@@ -1,6 +1,7 @@
 const fs = require('fs');
 const { cloudinary, userUploadOptions } = require('../config/cloudinaryConfig');
 const User = require('../models/User');
+const Comment = require('../models/Comment');
 
 exports.createUser = async (req, res) => {
   const {
@@ -129,7 +130,7 @@ exports.statusUser = async (req, res) => {
   const { status } = req.body;
 
   try {
-    const user = await User.findById( id );
+    const user = await User.findById(id);
     if (!user) throw new Error('the blog does not exist');
 
     user.status = status;
@@ -157,9 +158,11 @@ exports.deleteUser = async (req, res) => {
   const id = req.params.id;
 
   try {
+    if (id === '64c2d44f61cc7d6cec9d2abb') throw new Error('Dont remove admin');
     const user = await User.findByIdAndDelete(id);
     if (!user) throw new Error('the user does not exist');
     await cloudinary.uploader.destroy(user.id_image);
+    await Comment.deleteMany({ user_id: id });
 
     return res.status(200).json({ message: 'User has been deleted' });
   } catch (error) {
